# s2: Improve amd64 decompression speed.