NYC: 1566.5MB/s -> 1615.3MB/s (best of 10)

```
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkDecodeS2Block/0-html/block-32                               1455          1319          -9.35%
BenchmarkDecodeS2Block/0-html/block-better-32                        1698          1598          -5.89%
BenchmarkDecodeS2Block/0-html/block-best-32                          1638          1509          -7.88%
BenchmarkDecodeS2Block/1-urls/block-32                               17444         16722         -4.14%
BenchmarkDecodeS2Block/1-urls/block-better-32                        21008         20510         -2.37%
BenchmarkDecodeS2Block/1-urls/block-best-32                          22137         21418         -3.25%
BenchmarkDecodeS2Block/2-jpg/block-32                                144           148           +2.85%
BenchmarkDecodeS2Block/2-jpg/block-better-32                         143           148           +3.64%
BenchmarkDecodeS2Block/2-jpg/block-best-32                           150           156           +4.41%
BenchmarkDecodeS2Block/3-jpg_200b/block-32                           2.90          2.86          -1.38%
BenchmarkDecodeS2Block/3-jpg_200b/block-better-32                    3.41          3.40          -0.35%
BenchmarkDecodeS2Block/3-jpg_200b/block-best-32                      4.41          4.32          -2.00%
BenchmarkDecodeS2Block/4-pdf/block-32                                255           229           -10.08%
BenchmarkDecodeS2Block/4-pdf/block-better-32                         406           380           -6.40%
BenchmarkDecodeS2Block/4-pdf/block-best-32                           445           420           -5.64%
BenchmarkDecodeS2Block/5-html4/block-32                              9293          9067          -2.43%
BenchmarkDecodeS2Block/5-html4/block-better-32                       9453          9338          -1.22%
BenchmarkDecodeS2Block/5-html4/block-best-32                         9424          9338          -0.91%
BenchmarkDecodeS2Block/6-txt1/block-32                               4952          4891          -1.23%
BenchmarkDecodeS2Block/6-txt1/block-better-32                        6157          5963          -3.15%
BenchmarkDecodeS2Block/6-txt1/block-best-32                          5837          5783          -0.93%
BenchmarkDecodeS2Block/7-txt2/block-32                               3606          3574          -0.89%
BenchmarkDecodeS2Block/7-txt2/block-better-32                        5655          5416          -4.23%
BenchmarkDecodeS2Block/7-txt2/block-best-32                          5388          5403          +0.28%
BenchmarkDecodeS2Block/8-txt3/block-32                               14490         14472         -0.12%
BenchmarkDecodeS2Block/8-txt3/block-better-32                        16430         16164         -1.62%
BenchmarkDecodeS2Block/8-txt3/block-best-32                          16380         17107         +4.44%
BenchmarkDecodeS2Block/9-txt4/block-32                               16031         16072         +0.26%
BenchmarkDecodeS2Block/9-txt4/block-better-32                        23295         22180         -4.79%
BenchmarkDecodeS2Block/9-txt4/block-best-32                          23738         22471         -5.34%
BenchmarkDecodeS2Block/10-pb/block-32                                1444          1237          -14.34%
BenchmarkDecodeS2Block/10-pb/block-better-32                         1506          1302          -13.55%
BenchmarkDecodeS2Block/10-pb/block-best-32                           1343          1184          -11.84%
BenchmarkDecodeS2Block/11-gaviota/block-32                           6335          6033          -4.77%
BenchmarkDecodeS2Block/11-gaviota/block-better-32                    6196          5957          -3.86%
BenchmarkDecodeS2Block/11-gaviota/block-best-32                      6138          5895          -3.96%
BenchmarkDecodeS2Block/12-txt1_128b/block-32                         2.78          2.63          -5.19%
BenchmarkDecodeS2Block/12-txt1_128b/block-better-32                  2.84          2.79          -1.66%
BenchmarkDecodeS2Block/12-txt1_128b/block-best-32                    2.85          2.81          -1.61%
BenchmarkDecodeS2Block/13-txt1_1000b/block-32                        17.9          17.5          -2.07%
BenchmarkDecodeS2Block/13-txt1_1000b/block-better-32                 19.1          18.6          -2.57%
BenchmarkDecodeS2Block/13-txt1_1000b/block-best-32                   20.4          19.4          -4.86%
BenchmarkDecodeS2Block/14-txt1_10000b/block-32                       192           176           -8.68%
BenchmarkDecodeS2Block/14-txt1_10000b/block-better-32                264           244           -7.51%
BenchmarkDecodeS2Block/14-txt1_10000b/block-best-32                  273           246           -9.64%
BenchmarkDecodeS2Block/15-txt1_20000b/block-32                       332           320           -3.76%
BenchmarkDecodeS2Block/15-txt1_20000b/block-better-32                552           495           -10.30%
BenchmarkDecodeS2Block/15-txt1_20000b/block-best-32                  549           490           -10.87%
```